:construction: wip: Add attendance status
