+++ src/pages/Landing.js
@@ -27,8 +27,8 @@ const Landing = () => {
 >
   이동:{' '}
   <img src="/img/arrowkeys.png" alt="방향키" style={{ width: '40px' }} /> or{' '}
-  <img src="/img/mousekey.png" alt="방향키" style={{ width: '40px' }} /> / 점프:{' '}
-  <img src="/img/spacekey.png" alt="방향키" style={{ width: '50px' }} />
+  <img src="/img/mousekey.png" alt="마우스키" style={{ width: '40px' }} /> / 점프:{' '}
+  <img src="/img/spacekey.png" alt="스페이스바" style={{ width: '50px' }} />
 </p>
 
       <div
