--- src/styles/PetStatusAndAlarm.css
@@ -1,28 +1,0 @@
-.bottom-right {
-  width: 15vw;
-  height: 85vh;
-  display: flex;
-  flex-direction: column;
-  margin-right: 15px;
-}
-
-.notification,
-.status {
-  background-color: #ffffff;
-  border: 5px solid #d9d9d9;
-  padding: 10px 10px 10px 20px;
-  margin-bottom: 20px;
-  border-radius: 50px;
-  height: 40vh;
-  text-align: center;
-}
-
-.notification > h2,
-.status > h2 {
-  font-weight: lighter;
-  font-size: 24px;
-}
-
-.status p {
-  margin: 0;
-}
