+++ src/styles/Status.css
@@ -1,0 +1,87 @@
+.status-container{
+  width: 15vw;
+  height: 85vh;
+  display: flex;
+  flex-direction: column;
+  margin-right: 15px;
+}
+
+.attend,
+.status {
+  background-color: #ffffff;
+  border: 5px solid #d9d9d9;
+  padding: 10px 10px 10px 20px;
+  margin-bottom: 20px;
+  border-radius: 50px;
+  /* height: 40vh; */
+  text-align: center;
+}
+
+.attend > h2,
+.status > h2 {
+  font-weight: lighter;
+  font-size: 20px;
+  margin: 0;
+}
+
+.status p {
+  margin: 0;
+}
+
+.attend p {
+  font-weight: bold;
+  color: #F28434;
+  opacity: 0;
+  animation: slideUp 2s ease-in-out forwards;
+}
+
+.fire {
+  width: 50px;
+  animation: float 2s ease-in-out infinite;
+}
+
+.fire-sl{
+  position: relative;
+  top: 20px;
+  left: 30px;
+  width: 20px;
+  animation: float 2s ease-in-out infinite;
+}
+
+.fire-sr{
+  position: relative;
+  bottom: 30px;
+  right: 35px;
+  width: 20px;
+  animation: float 2s ease-in-out infinite;
+}
+
+@keyframes float {
+  0% {
+    transform: translateY(0);
+  }
+  50% {
+    transform: translateY(-10px);
+  }
+  100% {
+    transform: translateY(0);
+  }
+}
+@keyframes slideUp {
+  0% {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  30% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+  70% {
+    opacity: 1; /* 잠시 머무름 */
+    transform: translateY(0);
+  }
+  100% {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+}
